chore: moved arrow svgs to their own component
