+++ src/components/table.css
@@ -21,10 +21,11 @@
   height: fit-content;
   width: 100%;
   display: flex;
-  justify-content: flex-start;
+  justify-content: space-between;
 }
 .heading-arrow {
   padding-left: 5px;
+  float: right;
   display: inline-flex;
   align-items: center;
 }
